Only validate trees with modified files
Update ci/validate-manifests.sh so that by default it will only validate
overlays that contain files modified in the current commit. The '-a' flag
restores the original behavior of validating all discovered overlays.

This change makes the script more useful in a pre-push hook (where
otherwise the time required to build all the overlays can be annoying).